Add bootstrap modals on What I Currently Do page
Added in Zooniverse then added in bootstrap modal script. Identified a few
issues, such as modal appearing greyed out and freezing page, close button
not working, modal colour not working with the text colour and bootstrap
button css not lookin nice. To try to make the modal work, trialed two
stack overflow fixes and the second worked (adding some CSS in). Close
button then wasn't working to tried adding in and then removing some
javascript, and got it to work. Changed the bootstrap modal css so that
the background went from white to black so the light text could be read.